+++ sdk/src/main/scala/com/horizen/account/state/AccountStateView.scala
@@ -9,9 +9,9 @@ import com.horizen.consensus.ConsensusEpochNumber
 import com.horizen.evm.StateDB
 import com.horizen.state.StateView
 import com.horizen.utils.WithdrawalEpochInfo
+import scorex.util.{ModifierId, ScorexLogging}
 import sparkz.core.VersionTag
-import scorex.util.ScorexLogging
-import scorex.util.ModifierId
+
 import java.math.BigInteger
 
 // this class extends 2 main hierarchies, which are kept separate:
@@ -28,65 +28,12 @@ class AccountStateView(
     with AutoCloseable
     with ScorexLogging {
 
-
   def addTopQualityCertificates(refData: MainchainBlockReferenceData, blockId: ModifierId): Unit = {
     refData.topQualityCertificate.foreach(cert => {
       log.debug(s"adding top quality cert to state: $cert.")
       updateTopQualityCertificate(cert, blockId)
     })
   }
-
-  override def getAccountStorage(address: Array[Byte], key: Array[Byte]): Array[Byte] =
-    stateDb.getStorage(address, key)
-
-  override def updateAccountStorage(address: Array[Byte], key: Array[Byte], value: Array[Byte]): Unit =
-    stateDb.setStorage(address, key, value)
-
-  final override def removeAccountStorage(address: Array[Byte], key: Array[Byte]): Unit =
-    updateAccountStorage(address, key, null)
-
-  // random data used to salt chunk keys in the storage trie when accessed via get/updateAccountStorageBytes
-  private val chunkKeySalt =
-    BytesUtils.fromHexString("fa09428dd8121ea57327c9f21af74ffad8bfd5e6e39dc3dc6c53241a85ec5b0d")
-
-  // chunk keys are generated by hashing a salt, the original key and the chunk index
-  // the salt was added to reduce the risk of accidental hash collisions because similar strategies
-  // to generate storage keys might be used by the caller
-  private def getChunkKey(key: Array[Byte], chunkIndex: Int): Array[Byte] =
-    Keccak256.hash(chunkKeySalt, key, BigIntegerUtil.toUint256Bytes(BigInteger.valueOf(chunkIndex)))
-
-  final override def getAccountStorageBytes(address: Array[Byte], key: Array[Byte]): Array[Byte] = {
-    val length = new BigInteger(1, getAccountStorage(address, key)).intValueExact()
-    val data = new Array[Byte](length)
-    for (chunkIndex <- 0 until (length + Hash.LENGTH - 1) / Hash.LENGTH) {
-      getAccountStorage(address, getChunkKey(key, chunkIndex)).copyToArray(data, chunkIndex * Hash.LENGTH)
-    }
-    data
-  }
-
-  final override def updateAccountStorageBytes(address: Array[Byte], key: Array[Byte], value: Array[Byte]): Unit = {
-    // get previous length of value stored, if any
-    val oldLength = new BigInteger(1, getAccountStorage(address, key)).intValueExact()
-    // values are split up into 32-bytes chunks:
-    // the length of the value is stored at the original key and the chunks are stored at hash(key, i)
-    val newLength = value.length
-    // if the new value is empty remove all key-value pairs, including the one holding the value length
-    updateAccountStorage(address, key, BigIntegerUtil.toUint256Bytes(BigInteger.valueOf(newLength)))
-    for (start <- 0 until Math.max(newLength, oldLength) by Hash.LENGTH) {
-      val chunkIndex = start / Hash.LENGTH
-      val chunkKey = getChunkKey(key, chunkIndex)
-      if (start < newLength) {
-        // (over-)write chunks
-        updateAccountStorage(address, chunkKey, value.slice(start, start + Hash.LENGTH).padTo(Hash.LENGTH, 0.toByte))
-      } else {
-        // remove previous chunks that are not needed anymore
-        removeAccountStorage(address, chunkKey)
-      }
-    }
-  }
-
-  final override def removeAccountStorageBytes(address: Array[Byte], key: Array[Byte]): Unit =
-    updateAccountStorageBytes(address, key, Array.empty)
 
   // out-of-the-box helpers
   override def updateTopQualityCertificate(cert: WithdrawalEpochCertificate, blockId: ModifierId): Unit = {
@@ -124,8 +71,6 @@ class AccountStateView(
     metadataStorageView.commit(version)
   }
 
-  // getters
-
   override def getTopQualityCertificate(referencedWithdrawalEpoch: Int): Option[WithdrawalEpochCertificate] =
     metadataStorageView.getTopQualityCertificate(referencedWithdrawalEpoch)
 
@@ -142,7 +87,4 @@ class AccountStateView(
   }
 
   override def getAccountStateRoot: Array[Byte] = metadataStorageView.getAccountStateRoot
-
-  override def getGasTrackedView(gas: GasPool): BaseAccountStateView =
-    new AccountStateViewGasTracked(metadataStorageView, stateDb, messageProcessors, gas)
 }
